requirements.txt
PyYAML
numpy
Pillow
tensorflow~=2.0
tensorflow-hub~=0.6